Add a 'client-address' data expression
This is needed, at the very least, to fully support DHCPv6 prefix delegation
in the absence of a dynamic routing protocol (ie any homenet).  It allows an
`on commit` block to report the address to which the delegated prefix should
be routed, like so:

    on commit {
      execute("/usr/local/sbin/add-route-to-prefix",
              binary-to-ascii(16, 8, ":", option dhcp6.ia-pd),
              client-address,
              "eth0");
    }

The `add-route-to-prefix` script will need to parse the prefix out of the
encoded `ia-pd` data, but that's a Simple Matter of Programming.  The
client-address will come out exactly as needed for `ip route add`.