Supply dhcp6.ia-prefix to release/expiry event handlers
Although there's all sorts of information available to other parts of the
"release a lease" process, by the time the event handler is called, there's
no information available to do anything useful (like withdraw routes).  So,
I've hacked that in, via the least-offensive method I could think of (short
of massively re-engineering a *lot* of general-purpose code).

What I'm doing is setting the `dhcp6.ia-prefix` option to the IAPREFIX data
that's contained in the `struct iasubopt *lease` provided to
`move_lease_to_inactive`.  It's kinda-sorta lying, but not *really* badly (I
was careful to maintain the actual wire-format of a real IAPREFIX, just in
case).  This way, our release/expiry handlers can know which prefix is being
released, for route-withdrawal purposes.